// src/types/law.ts
export type LawNumType = {
  LawNum: Array<TextType>;
};

/**
 * 法令本体の型
 * @param LawBody 法令本体
 * @param ":@" 要素
 */
export type LawBodyType = {
  LawBody: Array<
    | LawTitleType
    | EnactStatementType
    | TOCType
    | PreambleType
    | MainProvisionType
    | SupplProvisionType
    | AppdxTableType
    | AppdxNoteType
    | AppdxStyleType
    | AppdxType
    | AppdxFigType
    | AppdxFormatType
  >;
  ":@": LawBodyAttributeType;
};

/**
 * 題名の型
 * @param LawTitle 題名
 * @param ":@" 要素
 */
export type LawTitleType = {
  LawTitle: Array<TextNodeType>;
  ":@": LawTitleAttributeType;
};

/**
 * 制定⽂の型
 * @param EnactStatement 制定⽂
 */
export type EnactStatementType = {
  EnactStatement: Array<TextNodeType>;
};

/**
 * ⽬次の型
 * @param TOC ⽬次
 */
export type TOCType = {
  TOC: Array<
    | TOCLabelType
    | TOCPreambleLabelType
    | TOCPartType
    | TOCChapterType
    | TOCSectionType
    | TOCArticleType
    | TOCSupplProvisionType
    | TOCAppdxTableLabelType
  >;
};

/**
 * ⽬次ラベルの型
 * @param TOCLabel ⽬次ラベル
 */
export type TOCLabelType = {
  TOCLabel: Array<TextNodeType>;
};

/**
 * ⽬次前⽂ラベルの型
 * @param TOCPreambleLabel ⽬次前⽂ラベル
 */
export type TOCPreambleLabelType = {
  TOCPreambleLabel: Array<TextNodeType>;
};

/**
 * ⽬次編の型
 * @param TOCPart ⽬次編
 * @param ":@" 要素
 */
export type TOCPartType = {
  TOCPart: Array<PartTitleType | ArticleRangeType | TOCChapterType>;
  ":@": CommonAttributeType;
};

/**
 * ⽬次章の型
 * @param TOCChapter ⽬次章
 * @param ":@" 要素
 */
export type TOCChapterType = {
  TOCChapter: Array<ChapterTitleType | ArticleRangeType | TOCSectionType>;
  ":@": CommonAttributeType;
};

/**
 * ⽬次節の型
 * @param TOCSection ⽬次節
 * @param ":@" 要素
 */
export type TOCSectionType = {
  TOCSection: Array<
    SectionTitleType | ArticleRangeType | TOCSubsectionType | TOCDivisionType
  >;
  ":@": CommonAttributeType;
};

/**
 * ⽬次款の型
 * @param TOCSection ⽬次款
 * @param ":@" 要素
 */
export type TOCSubsectionType = {
  TOCSubsection: Array<
    SubsectionTitleType | ArticleRangeType | TOCDivisionType
  >;
  ":@": CommonAttributeType;
};

/**
 * ⽬次款の型
 * @param TOCSection ⽬次款
 * @param ":@" 要素
 */
export type TOCDivisionType = {
  TOCDivision: Array<DivisionTitleType | ArticleRangeType>;
  ":@": CommonAttributeType;
};

/**
 * ⽬次条の型
 * @param TOCArticle ⽬次条
 * @param ":@" 要素
 */
export type TOCArticleType = {
  TOCArticle: Array<ArticleTitleType | ArticleCaptionType>;
  ":@": CommonAttributeType;
};

/**
 * ⽬次附則の型
 * @param TOCSupplProvision ⽬次附則
 */
export type TOCSupplProvisionType = {
  TOCSupplProvision: Array<
    SupplProvisionLabelType | ArticleRangeType | TOCArticleType | TOCChapterType
  >;
};

/**
 * ⽬次別表ラベルの型
 * @param TOCAppdxTableLabel ⽬次別表ラベル
 */
export type TOCAppdxTableLabelType = {
  TOCAppdxTableLabel: Array<TextNodeType>;
};

/**
 * 条範囲の型
 * @param ArticleRange 条範囲
 */
export type ArticleRangeType = {
  ArticleRange: Array<TextNodeType>;
};

/**
 * 前⽂の型
 * @param Preamble 前⽂
 */
export type PreambleType = {
  Preamble: Array<ParagraphType>;
};

/**
 * 本則の型
 * @param MainProvision 本則
 * @param ":@" 要素
 */
export type MainProvisionType = {
  MainProvision: Array<
    PartType | ChapterType | SectionType | ArticleType | ParagraphType
  >;
  ":@": MainProvisionAttributeType;
};

/**
 * 編の型
 * @param Part 編
 * @param ":@" 要素
 */
export type PartType = {
  Part: Array<PartTitleType | ArticleType | ChapterType>;
  ":@": CommonAttributeType;
};

/**
 * 編名の型
 * @param PartTitle 編名
 */
export type PartTitleType = {
  PartTitle: Array<TextNodeType>;
};

/**
 * 章の型
 * @param Chapter 章
 * @param ":@" 要素
 */
export type ChapterType = {
  Chapter: Array<ChapterTitleType | ArticleType | SectionType>;
  ":@": CommonAttributeType;
};

/**
 * 章名の型
 * @param ChapterTitle 章名
 */
export type ChapterTitleType = {
  ChapterTitle: Array<TextNodeType>;
};

/**
 * 節の型
 * @param Section 節
 * @param ":@" 要素
 */
export type SectionType = {
  Section: Array<
    SectionTitleType | ArticleType | SubsectionType | DivisionType
  >;
  ":@": CommonAttributeType;
};

/**
 * 節名の型
 * @param SectionTitle 節名
 */
export type SectionTitleType = {
  SectionTitle: Array<TextNodeType>;
};

/**
 * 款の型
 * @param Subsection 款
 * @param ":@" 要素
 */
export type SubsectionType = {
  Subsection: Array<SubsectionTitleType | ArticleType | DivisionType>;
  ":@": CommonAttributeType;
};

/**
 * 款名の型
 * @param SubsectionTitle 款名
 */
export type SubsectionTitleType = {
  SubsectionTitle: Array<TextNodeType>;
};

/**
 * ⽬の型
 * @param Division ⽬
 * @param ":@" 要素
 */
export type DivisionType = {
  Division: Array<DivisionTitleType | ArticleType>;
  ":@": CommonAttributeType;
};

/**
 * ⽬名の型
 * @param DivisionTitle ⽬名
 */
export type DivisionTitleType = {
  DivisionTitle: Array<TextNodeType>;
};

/**
 * 条の型
 * @param Article 条
 * @param ":@" 要素
 */
export type ArticleType = {
  Article: Array<
    ArticleCaptionType | ArticleTitleType | ParagraphType | SupplNoteType
  >;
  ":@": CommonAttributeType;
};

/**
 * 条名の型
 * @param ArticleTitle 条名
 */
export type ArticleTitleType = {
  ArticleTitle: Array<TextNodeType>;
};

/**
 * 条⾒出しの型
 * @param Article 条⾒出し
 * @param ":@" 要素
 */
export type ArticleCaptionType = {
  ArticleCaption: Array<TextNodeType>;
  ":@": CaptionAttributeType;
};

/**
 * 項の型
 * @param Paragraph 項
 * @param ":@" 要素
 */
export type ParagraphType = {
  Paragraph: Array<
    | ParagraphCaptionType
    | ParagraphNumType
    | ParagraphSentenceType
    | AmendProvisionType
    | ClassType
    | ItemType
    | TableStructType
    | FigStructType
    | StyleStructType
    | ListType
  >;
  ":@": ParagraphAttributeType;
};

/**
 * 項⾒出しの型
 * @param ParagraphCaption 項⾒出し
 */
export type ParagraphCaptionType = {
  ParagraphCaption: Array<TextNodeType>;
  ":@": CaptionAttributeType;
};

/**
 * 項番号の型
 * @param ParagraphNum 項番号
 */
export type ParagraphNumType = {
  ParagraphNum: Array<TextNodeType>;
};

/**
 * 項⽂の型
 * @param ParagraphSentence 項⽂
 */
export type ParagraphSentenceType = {
  ParagraphSentence: Array<SentenceType>;
};

/**
 * 付記の型
 * @param SupplNote 付記
 */
export type SupplNoteType = {
  SupplNote: Array<TextNodeType>;
};

/**
 * 改正規定の型
 * @param AmendProvision 改正規定
 */
export type AmendProvisionType = {
  AmendProvision: Array<AmendProvisionSentenceType | NewProvisionType>;
};

/**
 * 改正規定⽂の型
 * @param AmendProvisionSentence 改正規定⽂
 */
export type AmendProvisionSentenceType = {
  AmendProvisionSentence: Array<SentenceType>;
};

/**
 * 新規定の型
 * @param NewProvision 新規定
 */
export type NewProvisionType = {
  NewProvision: Array<
    | LawTitleType
    | PreambleType
    | TOCType
    | PartType
    | PartTitleType
    | ChapterType
    | ChapterTitleType
    | SectionType
    | SectionTitleType
    | SubsectionType
    | SubsectionTitleType
    | DivisionType
    | DivisionTitleType
    | ArticleType
    | SupplNoteType
    | ParagraphType
    | ItemType
    | Subitem1Type
    | Subitem2Type
    | Subitem3Type
    | Subitem4Type
    | Subitem5Type
    | Subitem6Type
    | Subitem7Type
    | Subitem8Type
    | Subitem9Type
    | Subitem10Type
    | ListType
    | SentenceType
    | AmendProvisionType
    | AppdxTableType
    | AppdxNoteType
    | AppdxStyleType
    | AppdxType
    | AppdxFigType
    | AppdxFormatType
    | SupplProvisionAppdxStyleType
    | SupplProvisionAppdxTableType
    | SupplProvisionAppdxType
    | TableStructType
    | TableRowType
    | TableColumnType
    | FigStructType
    | NoteStructType
    | StyleStructType
    | FormatStructType
    | RemarksType
    | LawBodyType
  >;
};

/**
 * 類の型
 * @param Class 類
 * @param ":@" 要素
 */
export type ClassType = {
  Class: Array<ClassTitleType | ClassSentenceType | ItemType>;
  ":@": CommonAttributeType;
};

/**
 * 類名の型
 * @param ClassTitle 類名
 */
export type ClassTitleType = {
  ClassTitle: Array<TextNodeType>;
};

/**
 * 類⽂の型
 * @param ClassSentence 類⽂
 */
export type ClassSentenceType = {
  ClassSentence: Array<SentenceType | ColumnType | TableType>;
};

/**
 * 号の型
 * @param Item 号
 * @param ":@" 要素
 */
export type ItemType = {
  Item: Array<
    | ItemTitleType
    | ItemSentenceType
    | Subitem1Type
    | TableStructType
    | FigStructType
    | StyleStructType
    | ListType
  >;
  ":@": CommonAttributeType;
};

/**
 * 号名の型
 * @param ItemTitle 号名
 */
export type ItemTitleType = {
  ItemTitle: Array<TextNodeType>;
};

/**
 * 号⽂の型
 * @param ItemSentence 号⽂
 */
export type ItemSentenceType = {
  ItemSentence: Array<SentenceType | ColumnType | TableType>;
};

/**
 * 号細分の型
 * @param Subitem1 号細分
 * @param ":@" 要素
 */
export type Subitem1Type = {
  Subitem1: Array<
    | Subitem1TitleType
    | Subitem1SentenceType
    | Subitem2Type
    | TableStructType
    | FigStructType
    | StyleStructType
    | ListType
  >;
  ":@": CommonAttributeType;
};

/**
 * 号細分名の型
 * @param Subitem1Title 号細分名
 */
export type Subitem1TitleType = {
  Subitem1Title: Array<TextNodeType>;
};

/**
 * 号細分⽂の型
 * @param Subitem1Sentence 号細分⽂
 */
export type Subitem1SentenceType = {
  Subitem1Sentence: Array<SentenceType | ColumnType | TableType>;
};

/**
 * 号細分２の型
 * @param Subitem2 号細分２
 * @param ":@" 要素
 */
export type Subitem2Type = {
  Subitem2: Array<
    | Subitem2TitleType
    | Subitem2SentenceType
    | Subitem3Type
    | TableStructType
    | FigStructType
    | StyleStructType
    | ListType
  >;
  ":@": CommonAttributeType;
};

/**
 * 号細分２名の型
 * @param Subitem2Title 号細分２名
 */
export type Subitem2TitleType = {
  Subitem2Title: Array<TextNodeType>;
};

/**
 * 号細分２⽂の型
 * @param Subitem2Sentence 号細分２⽂
 */
export type Subitem2SentenceType = {
  Subitem2Sentence: Array<SentenceType | ColumnType | TableType>;
};

/**
 * 号細分３の型
 * @param Subitem3 号細分３
 * @param ":@" 要素
 */
export type Subitem3Type = {
  Subitem3: Array<
    | Subitem3TitleType
    | Subitem3SentenceType
    | Subitem4Type
    | TableStructType
    | FigStructType
    | StyleStructType
    | ListType
  >;
  ":@": CommonAttributeType;
};

/**
 * 号細分３名の型
 * @param Subitem3Title 号細分３名
 */
export type Subitem3TitleType = {
  Subitem3Title: Array<TextNodeType>;
};

/**
 * 号細分３⽂の型
 * @param Subitem3Sentence 号細分３⽂
 */
export type Subitem3SentenceType = {
  Subitem3Sentence: Array<SentenceType | ColumnType | TableType>;
};

/**
 * 号細分４の型
 * @param Subitem4 号細分４
 * @param ":@" 要素
 */
export type Subitem4Type = {
  Subitem4: Array<
    | Subitem4TitleType
    | Subitem4SentenceType
    | Subitem5Type
    | TableStructType
    | FigStructType
    | StyleStructType
    | ListType
  >;
  ":@": CommonAttributeType;
};

/**
 * 号細分４名の型
 * @param Subitem4Title 号細分４名
 */
export type Subitem4TitleType = {
  Subitem4Title: Array<TextNodeType>;
};

/**
 * 号細分４⽂の型
 * @param Subitem4Sentence 号細分４⽂
 */
export type Subitem4SentenceType = {
  Subitem4Sentence: Array<SentenceType | ColumnType | TableType>;
};

/**
 * 号細分５の型
 * @param Subitem5 号細分５
 * @param ":@" 要素
 */
export type Subitem5Type = {
  Subitem5: Array<
    | Subitem5TitleType
    | Subitem5SentenceType
    | Subitem6Type
    | TableStructType
    | FigStructType
    | StyleStructType
    | ListType
  >;
  ":@": CommonAttributeType;
};

/**
 * 号細分５名の型
 * @param Subitem5Title 号細分５名
 */
export type Subitem5TitleType = {
  Subitem5Title: Array<TextNodeType>;
};

/**
 * 号細分５⽂の型
 * @param Subitem5Sentence 号細分５⽂
 */
export type Subitem5SentenceType = {
  Subitem5Sentence: Array<SentenceType | ColumnType | TableType>;
};

/**
 * 号細分６の型
 * @param Subitem6 号細分６
 * @param ":@" 要素
 */
export type Subitem6Type = {
  Subitem6: Array<
    | Subitem6TitleType
    | Subitem6SentenceType
    | Subitem7Type
    | TableStructType
    | FigStructType
    | StyleStructType
    | ListType
  >;
  ":@": CommonAttributeType;
};
// ...
export type ArithFormulaType = {
  ArithFormula: Array<LawTypeList>;
  ":@": NumAttributeType;
};
// ...
export type QuoteStructType = { QuoteStruct: Array<LawTypeList> };

/**
 * ルビ構造の型
 * @param Ruby ルビ構造
 */
export type RubyType = {
  Ruby: Array<TextType | RtType>;
};

/**
 * ルビの型
 * @param Rt ルビ
 */
export type RtType = {
  Rt: Array<TextType>;
};

/**
 * 傍線の型
 * @param Line 傍線
 * @param Style 線種
 */
export type LineType = {
  Line: Array<TextNodeType>;
  Style: BorderStyleType;
};

/**
 * 上付き⽂字の型
 * @param Sup 上付き⽂字
 */
export type SupType = { Sup: Array<TextType> };

/**
 * 下付き⽂字の型
 * @param Sub 下付き⽂字
 */
export type SubType = { Sub: Array<TextType> };

/**
 * テキストの型
 * @param _ テキスト
 */
export type TextType = {
  _: string;
};

/**
 * テキストノードの型
 */
export type TextNodeType =
  | LineType
  | RubyType
  | SupType
  | SubType
  | TextType
  | QuoteStructType
  | ArithFormulaType;

/**
 * 法令に関する型の一覧
 */
export type LawTypeList =
  | LawNumType
  | LawBodyType
  | LawTitleType
  | EnactStatementType
  | TOCType
  | TOCLabelType
  | TOCPreambleLabelType
  | TOCPartType
  | TOCChapterType
  | TOCSectionType
  | TOCSubsectionType
  | TOCDivisionType
  | TOCArticleType
  | TOCSupplProvisionType
  | TOCAppdxTableLabelType
  | ArticleRangeType
  | PreambleType
  | MainProvisionType
  | PartType
  | PartTitleType
  | ChapterType
  | ChapterTitleType
  | SectionType
  | SectionTitleType
  | SubsectionType
  | SubsectionTitleType
  | DivisionType
  | DivisionTitleType
  | ArticleType
  | ArticleTitleType
  | ArticleCaptionType
  | ParagraphType
  | ParagraphCaptionType
  | ParagraphNumType
  | ParagraphSentenceType
  | SupplNoteType
  | AmendProvisionType
  | AmendProvisionSentenceType
  | NewProvisionType
  | ClassType
  | ClassTitleType
  | ClassSentenceType
  | ItemType
  | ItemTitleType
  | ItemSentenceType
  | Subitem1Type
  | Subitem1TitleType
  | Subitem1SentenceType
  | Subitem2Type
  | Subitem2TitleType
  | Subitem2SentenceType
  | Subitem3Type
  | Subitem3TitleType
  | Subitem3SentenceType
  | Subitem4Type
  | Subitem4TitleType
  | Subitem4SentenceType
  | Subitem5Type
  | Subitem5TitleType
  | Subitem5SentenceType
  | Subitem6Type
  | Subitem6TitleType
  | Subitem6SentenceType
  | Subitem7Type
  | Subitem7TitleType
  | Subitem7SentenceType
  | Subitem8Type
  | Subitem8TitleType
  | Subitem8SentenceType
  | Subitem9Type
  | Subitem9TitleType
  | Subitem9SentenceType
  | Subitem10Type
  | Subitem10TitleType
  | Subitem10SentenceType
  | SentenceType
  | ColumnType
  | SupplProvisionType
  | SupplProvisionLabelType
  | SupplProvisionAppdxTableType
  | SupplProvisionAppdxTableTitleType
  | SupplProvisionAppdxStyleType
  | SupplProvisionAppdxStyleTitleType
  | SupplProvisionAppdxType
  | AppdxTableType
  | AppdxTableTitleType
  | AppdxNoteType
  | AppdxNoteTitleType
  | AppdxStyleType
  | AppdxStyleTitleType
  | AppdxFormatType
  | AppdxFormatTitleType
  | AppdxType
  | ArithFormulaNumType
  | ArithFormulaType
  | AppdxFigType
  | AppdxFigTitleType
  | TableStructType
  | TableStructTitleType
  | TableType
  | TableRowType
  | TableHeaderRowType
  | TableHeaderColumnType
  | TableColumnType
  | FigStructType
  | FigStructTitleType
  | FigType
  | NoteStructType
  | NoteStructTitleType
  | NoteType
  | StyleStructType
  | StyleStructTitleType
  | StyleType
  | FormatStructType
  | FormatStructTitleType
  | FormatType
  | RelatedArticleNumType
  | RemarksType
  | RemarksLabelType
  | ListType
  | ListSentenceType
  | Sublist1Type
  | Sublist1SentenceType
  | Sublist2Type
  | Sublist2SentenceType
  | Sublist3Type
  | Sublist3SentenceType
  | QuoteStructType
  | RubyType
  | LineType
  | SupType
  | SubType
  | TextType;
// ...
export type NumAttributeType = {
  Num?: number;
};
// ...
export type BorderStyleType = (typeof BorderStyle)[keyof typeof BorderStyle];